+++ frontend/wailsjs/go/main/App.d.ts
@@ -36,6 +36,8 @@ export function ListRemovableDrives():Promise<main.ListDrivesResult>;
 
 export function Manufacturers():Promise<Array<string>>;
 
+export function Models(arg1:string,arg2:string):Promise<Array<string>>;
+
 export function NewCsvTemplate():Promise<main.PathResult>;
 
 export function OpenConfiguration():Promise<main.OpenConfigResult>;