+++ frontend/wailsjs/go/main/App.js
@@ -66,6 +66,10 @@ export function Manufacturers() {
   return window['go']['main']['App']['Manufacturers']();
 }
 
+export function Models(arg1, arg2) {
+  return window['go']['main']['App']['Models'](arg1, arg2);
+}
+
 export function NewCsvTemplate() {
   return window['go']['main']['App']['NewCsvTemplate']();
 }